scripts: extract_dts_include: Generate pinmux per instance
When several instances of pinmux are present (NXP case),
generate an array of enabled instances and a pinmux array
per instance.
Add a label in dts for pinmux in this case.
Add a 'instance' key in in pinctrl key in structs dict to
hold this information.
Tested label does not harm if only one pinmux instance (stm32 case)

Signed-off-by: Erwan Gouriou <erwan.gouriou@linaro.org>